Added classes to be able to parse the locations of DRIPs from XML, a double should be enough for coordinates, added HasLocation since a DRIP could lack a location due to bad documentation at roadmanagers and uuid/guid is not correct for a unique sign since in Belgium it is basically a string.

+++ FileModels/LaneControlSignalLocations/Geometry.cs
@@ -6,6 +6,6 @@ namespace Matrix.FileModels.LaneControlSignalLocations
     {
         public string Type { get; set; }
 
-        public IEnumerable<float> Coordinates { get; set; }
+        public IEnumerable<double> Coordinates { get; set; }
     }
 }
